# sdks/sdkwork-models-backend-sdk/openapi/sdkwork-models-backend-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Models Backend API",
    "version": "0.1.0",
    "description": "Intelligence catalog API owned by sdkwork-models. Materialized from composed host catalog mount OpenAPI authority."
  },
  "servers": [
    {
      "url": "/backend/v3/api"
    }
  ],
  "paths": {
...
    "/backend/v3/api/ai/models": {
      "get": {
        "description": "List. Reads ai. Writes none. File targets none.",
        "operationId": "models.list",
        "parameters": [
          {
            "description": "One-based page index.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Number of models returned in the current page.",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Free-text model search query.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Billing meter code filter.",
            "in": "query",
            "name": "billing_meter",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated vendor code filter.",
            "explode": false,
            "in": "query",
            "name": "vendor_codes",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "style": "form"
          },
          {
            "description": "Comma-separated modality filter.",
            "explode": false,
            "in": "query",
            "name": "modalities",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "style": "form"
          },
          {
            "description": "Comma-separated capability filter.",
            "explode": false,
            "in": "query",
            "name": "capabilities",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "style": "form"
          },
          {
            "description": "Comma-separated model category filter.",
            "explode": false,
            "in": "query",
            "name": "categories",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "style": "form"
          },
          {
            "description": "Comma-separated channel group filter.",
            "explode": false,
            "in": "query",
            "name": "groups",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "style": "form"
          },
          {
            "description": "Comma-separated model type labels (Chat, Image, Embedding, ...).",
            "in": "query",
            "name": "model_types",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsListResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Server Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Error response."
          }
        },
        "security": [
          {
            "AccessToken": [],
            "AuthToken": []
          }
        ],
        "summary": "List",
        "tags": [
          "ai"
        ],
        "x-contract-kind": "read",
        "x-file-targets": [],
        "x-read-sources": [
          "ai"
        ],
        "x-route-scope": "admin",
        "x-sdk-domain": "intelligence",
        "x-sdkwork-api-authority": "sdkwork-models-backend-api",
        "x-sdkwork-domain": "intelligence",
        "x-sdkwork-owner": "sdkwork-models",
        "x-sdkwork-resource": "models",
        "x-sdkwork-source-route-crate": "sdkwork-routes-models-catalog-backend-api",
        "x-source-file": "tools/bootstrap_frontend_contract_from_route_manifest.py",
        "x-write-tables": []
      },
      "post": {
        "description": "Create. Reads ai. Writes ops_audit_log. File targets none.",
        "operationId": "models.create",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsCreateResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Server Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Error response."
          }
        },
        "security": [
          {
            "AccessToken": [],
            "AuthToken": []
          }
        ],
        "summary": "Create",
        "tags": [
          "ai"
        ],
        "x-contract-kind": "create",
        "x-file-targets": [],
        "x-read-sources": [
          "ai"
        ],
        "x-route-scope": "admin",
        "x-sdk-domain": "intelligence",
        "x-sdkwork-api-authority": "sdkwork-models-backend-api",
        "x-sdkwork-domain": "intelligence",
        "x-sdkwork-owner": "sdkwork-models",
        "x-sdkwork-resource": "models",
        "x-sdkwork-source-route-crate": "sdkwork-routes-models-catalog-backend-api",
        "x-source-file": "tools/bootstrap_frontend_contract_from_route_manifest.py",
        "x-write-tables": [
          "ops_audit_log"
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AccessToken": {
        "in": "header",
        "name": "Access-Token",
        "type": "apiKey"
      },
      "AuthToken": {
        "bearerFormat": "SDKWork auth token",
        "scheme": "bearer",
        "type": "http"
      }
    },
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 40011,
            "maximum": 40099
          },
          "i18nKey": {
            "type": "string"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                }
              ]
            }
          }
        }
      },
      "JsonNull": {
        "description": "JSON null value.",
        "type": "null"
      },
      "JsonObject": {
        "additionalProperties": {
          "allOf": [
            {
              "$ref": "#/components/schemas/JsonValue"
            }
          ],
          "description": "Schema fragment used by json object."
        },
        "description": "JSON object with typed JSON values.",
        "type": "object"
      },
      "JsonValue": {
        "description": "JSON value accepted by flexible Claw Router metadata and extension maps.",
        "oneOf": [
          {
            "description": "String JSON value.",
            "type": "string"
          },
          {
            "description": "Number JSON value.",
            "type": "number"
          },
          {
            "description": "Integer JSON value.",
            "type": "integer"
          },
          {
            "description": "Boolean JSON value.",
            "type": "boolean"
          },
          {
            "description": "Array JSON value.",
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "array"
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/JsonObject"
              }
            ],
            "description": "Object JSON value."
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/JsonNull"
              }
            ],
            "description": "Null JSON value."
          }
        ]
      },
      "ModelCatalogPage": {
        "additionalProperties": false,
        "description": "Model catalog page schema exposed by Claw Router.",
        "name": "ModelCatalogPage",
        "properties": {
          "groups": {
            "description": "Groups field on model catalog page.",
            "items": {
              "additionalProperties": false,
              "description": "Groups field on model catalog page.",
              "properties": {
                "key": {
                  "description": "Key field on model catalog page.",
                  "type": "string"
                },
                "label": {
                  "description": "Label field on model catalog page.",
                  "type": "string"
                },
                "modelCount": {
                  "description": "Model count field on model catalog page.",
                  "format": "int64",
                  "pattern": "^[0-9]+$",
                  "type": "string",
                  "x-sdkwork-int64-string": true,
                  "x-sdkwork-rust-type": "i64"
                }
              },
              "required": [
                "key",
                "label",
                "modelCount"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "items": {
            "description": "Items field on model catalog page.",
            "items": {
              "additionalProperties": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/JsonValue"
                  }
                ],
                "description": "Items field on model catalog page."
              },
              "description": "Items field on model catalog page.",
              "type": "object"
            },
            "type": "array"
          },
          "pageInfo": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PageInfo"
              }
            ],
            "description": "Page info field on model catalog page."
          }
        },
        "required": [
          "items",
          "groups",
          "pageInfo"
        ],
        "type": "object"
      },
...
      "ModelsCreateResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "additionalProperties": false,
            "description": "Object schema used by models create result.",
            "properties": {
              "data": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/NoData"
                  }
                ],
                "description": "No business data returned by this operation."
              }
            },
            "required": [
              "data"
            ],
            "type": "object"
          }
        ],
        "description": "Models create result schema exposed by Claw Router.",
        "x-operation-id": "models.create"
      },
...
      "ModelsListResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "additionalProperties": false,
            "description": "Object schema used by models list result.",
            "properties": {
              "data": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ModelCatalogPage"
                  }
                ],
                "description": "Data field on models list result."
              }
            },
            "required": [
              "data"
            ],
            "type": "object"
          }
        ],
        "description": "Models list result schema exposed by Claw Router.",
        "x-operation-id": "models.list"
      },
...
      "NoData": {
        "additionalProperties": false,
        "description": "Closed empty payload for operations that complete without business data.",
        "properties": {},
        "type": "object"
      },
      "PageInfo": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "offset",
              "cursor"
            ]
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "totalItems": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "i18nKey": {
            "type": "string",
            "description": "Optional stable localization key such as errors.result.40001."
          },
          "locale": {
            "type": "string",
            "description": "Optional effective BCP 47 locale used by framework message mapping."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md §15.3."
      },
...
    }
  }
}